# Conflicts:
#	api/include/opentelemetry/trace/trace_state.h
#	api/test/trace/trace_state_test.cc

+++ ext/test/zpages/threadsafe_span_data_test.cc
@@ -52,5 +52,5 @@ TEST(ThreadsafeSpanData, Set)
   ASSERT_EQ(data.GetDescription(), "description");
   ASSERT_EQ(data.GetStartTime().time_since_epoch(), now.time_since_epoch());
   ASSERT_EQ(data.GetDuration(), std::chrono::nanoseconds(1000000));
-  ASSERT_EQ(opentelemetry::nostd::get<int64_t>(data.GetAttributes().at("attr1")), 314159);
+  ASSERT_EQ(opentelemetry::nostd::get<int>(data.GetAttributes().at("attr1")), 314159);
 }
